Alot of stuff.
Removing NodeAttributes member from the DARSEvent class and replacing it with int and string fields. Doing this to make serialization to file easier and remain consistent. 

Not using the connections class for right now. Just keep it around for later.

Removing printfs from gui. 

Making the NodeAttributes area use the node inspector interface and listen for when a node is selected.

Getting rid of the "spawn a hunded nodes" code that ran when you start the app.

Fixing a null pointer exception that I ran across in testing from NodeStore.java.

Removing the "atts" member from the proto interface. It's up to the implementors of the proto to decide if they should have it or not. The interfaces of getNodeAttributes and setNodeAttributes are sufficient.

Reworking the NodeAttributes member of DARSEvent. Unfortunately this meant some changes had to be done to the sim engine code. They are superficial though.

Finally, adding an ID assigning routine to the sim engine.

At this current revision, we can now add nodes from the gui down to the simulator, have the simulator assign a new node, and then push it back up into the gui.