Drastically refactoring gui.java. Creating a new class called DARSAppMenu. This class holds all the logic and initialization for the menu bar and the action panel at the top of the screen. Misael will be managing this component.
GUI.java should be thought of as an executive class. It should only be directing actions to the gui components, mostly from consumeInput. Let's try to keep it as clean as possible. 

Also, redoing some things, particularly having to with the CLEAR_SIM and adding a NEW_SIM event. I didn't see a reason for the simEngine to invoke InputHandler itself to clear out all the nodes. This would result in multiple IN/OUT "DELETE_NODE" events to be generated for no real reason. Instead, the sim area now has a clear() method and the node store has a clear method (which just calls the clear() method of the java HashMap.