Reworking the node inspector interface. I believe that we no longer need to concern ourselves with concurrency with regards to this module (updating watch dialogs, creating dialogs, viewing node attributes, etc).
The implementer of the node inspector interface is now the sim engine, before it was the node store. It makes sense to implement it through the sim engine because the sim engine is the manager of the simulation thread. Now the consumeInput(), mainLoop(), and the node inspector interface all share the same lock object (mutex). If this becomes inefficient we can granularize the concurrency better.