Adding working replays. Added a parseReplay function to parser which parses all the replay events into memory.
Added a Replayer class too. I chose to make a Replayer implement the OutputConsumer interface and make it only care about OUT_QUANTUM_ELAPSED events. This way, the speed bar at the top of the GUI works for replays too. 

There is still much work to be done in the way of locking down the GUI during replays..