Refactoring node attributes to make it an immutable data type. This way, it's safe to share it among threads (we don't have to re-allocate a new instance every call to node.getAttributes() ).
From henceforth, whenever we are dealing with a node attributes object, it's guaranteed to be complete and un-modifyable. This ends up making updates on a single field of the data type rather expensive, but reads are cheap, and they vastly outnumber writes. 

Also adding a Defaults class as a place to store application default constants.