Refactoring simEngine a bit to make Start_Sim have a simspeed argument as opposed to generating a separate simSpeed event immediately afterward.
Adding support for dynamic sim types on the GUI. The menu items and popup dialogs that ask the user what type of sim they'd like to run are dynamically populated based on the enums in DARSEvent.SimType. 

This is a really sweet feature because someone can now implement an entirely new protocol without having to change a thing in the GUI. They would just need to add an enum to DARSEvent.SimType.