A couple bug fixes:
Fixing bug where nodes were not being checked for "selected"ness when deleting from the simArea.  (Kenny's bug, occurs when you delete the last node on the screen using the delete key.) 

It turns out that when a node is deleted, the nodeAttributesArea removes an entry from its node combobox. This fires a swing change event. This event is processed by the combobox's action handler that ultimately does a simArea.select(). I left this in because it seems intuitive as a feature. 

Next, the logArea was changed to prepend a timestamp for every log (and got rid of the old code for timestamps in DARSEvent).

Finally, the clear action was changed to not stop the simulator and to not reset the quantum counter display. I have no idea what the intention was here for the clear action, but as of this writing the simulator and the GUI did not treat them the same way. 

I changed it to just generate a clear_sim event, which is handled in the following way:
1. Remove all nodes (on screen and under the hood).
2. Truncate the log area.