Making log messages more descriptive by dividing them into three classes:
NODE INFO:  Log message that informs the user of something that a node did.
 
SIM INFO: Log message that informs the user of something that the simulator did.

SIM_ERROR: A real error happened. A Logic error or some other programming level error.

In support of this, there's a new DARS event. outNodeInfo event. Protocols can use this a generic way to inform the user of an important action taking place in the protocol (like the TTL of a narrative message expiring, for instance).